Dev branch

+++ app/Http/Controllers/CampaignGamePlayController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\CampaignGamePlay;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\GamePlayRequest;
 use App\Services\CampaignGamePlay\CampaignGamePlayService;
 use App\Http\Requests\Campaign\UpdateCampaignGamePlayRequest;
@@ -20,6 +21,7 @@ class CampaignGamePlayController extends BaseController
     
     public function storeCampaignGamePlay(GamePlayRequest $request, $campaignId, $gameId) {
         try {
+            Gate::authorize('is-audience');
             $data =  $this->campaignGamePlayService->storeCampaignGamePlay($request, $campaignId, $gameId);
     
         } catch (\Exception $e){
@@ -32,6 +34,7 @@ class CampaignGamePlayController extends BaseController
     public function index()
     {
         try {
+            Gate::authorize('is-audience');
             return response()->json(CampaignGamePlay::with('game', 'user', 'campaign')->get());
 
         } catch (\Throwable $th) {
@@ -47,6 +50,7 @@ class CampaignGamePlayController extends BaseController
     public function show($campaignId, $gameId)
     {
         try {
+            Gate::authorize('is-audience');
             $data =  $this->campaignGamePlayService->show($campaignId, $gameId);
     
         } catch (\Exception $e){
@@ -59,7 +63,8 @@ class CampaignGamePlayController extends BaseController
     public function update(UpdateCampaignGamePlayRequest $request, $campaignId, $gameId)
     {
         try {
-         $data =  $this->campaignGamePlayService->update($request, $campaignId, $gameId);
+            Gate::authorize('is-audience');
+            $data =  $this->campaignGamePlayService->update($request, $campaignId, $gameId);
     
         } catch (\Exception $e){
             return $this->sendError("something went wrong", ['error' => $e->getMessage()], 500);
@@ -71,7 +76,8 @@ class CampaignGamePlayController extends BaseController
     public function destroy($campaignId, $gameId)
     {
         try {
-         $data =  $this->campaignGamePlayService->destroy($campaignId, $gameId);
+            Gate::authorize('is-user');
+            $data =  $this->campaignGamePlayService->destroy($campaignId, $gameId);
     
         } catch (\Exception $e){
             return $this->sendError("something went wrong", ['error' => $e->getMessage()], 500);
